cpu_input_boost: Allow configuration of each cluster's minimum idle frequency
Because we will be using get_min_freq to raise default minimums to
higher, more efficient frequencies, add a secondary function to
return stock minimums when the device is idle.

In this way, we can balance the utilization of more performant
frequencies while the device is in use while retaining the ability
to scale down to hardware minimums when it isn't.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Little-W <1405481963@qq.com>